+++ app/services/adal/adal-config.ts
@@ -1,0 +1,4 @@
+export interface AdalConfig {
+    tenant: string;
+    clientId: string;
+}
