+++ app/services/adal/adal-constants.ts
@@ -1,0 +1,24 @@
+export const baseUrl = "https://login.microsoftonline.com";
+export const defaultTenant = "common";
+
+export interface AuthorizeUrlParams {
+    response_type: string;
+    redirect_uri: string;
+    client_id: string;
+    scope: string;
+    nonce?: string;
+    state?: string;
+    resource: string;
+    prompt?: string;
+}
+
+export function authorizeUrl(tenant: string, params: AuthorizeUrlParams) {
+    const query = objectToParams(params);
+    return `${baseUrl}/${tenant}/oauth2/authorize?${query}`;
+}
+
+export function objectToParams(object): string {
+    return Object.keys(object).map((key) => {
+        return `${encodeURIComponent(key)}=${object[key]}`;
+    }).join("&");
+}
