+++ app/services/adal/index.ts
@@ -1,0 +1,1 @@
+export * from "./adal.service";
