+++ app/services/index.ts
@@ -5,7 +5,7 @@ export * from "./pool-service";
 export * from "./node-service";
 export * from "./settings-service";
 export * from "./task-service";
-export * from "./adal-service";
+export * from "./adal";
 
 // This needs to be last(as it does dynamic inject which problably have dependencies on above services)
 export * from "./command-service";
